+++ Source/Libraries/CorruptCore/CorruptCore_Extensions.cs
@@ -30,153 +30,6 @@ namespace RTCV.CorruptCore
             }
         }
 
-        #region STRING EXTENSIONS
-
-        //taken from https://stackoverflow.com/questions/11454004/calculate-a-md5-hash-from-a-string
-        public static string CreateMD5(this string input)
-        {
-            // Use input string to calculate MD5 hash
-            using (var md5 = System.Security.Cryptography.MD5.Create())
-            {
-                var inputBytes = System.Text.Encoding.ASCII.GetBytes(input);
-                var hashBytes = md5.ComputeHash(inputBytes);
-
-                // Convert the byte array to hexadecimal string
-                var sb = new StringBuilder();
-                for (var i = 0; i < hashBytes.Length; i++)
-                {
-                    sb.Append(hashBytes[i].ToString("X2"));
-                }
-                return sb.ToString();
-            }
-        }
-
-        /// <summary>
-        /// Gets you a byte array from the CONTENTS of a string
-        /// </summary>
-        /// <param name="hex"></param>
-        /// <returns></returns>
-        public static byte[] StringToByteArray(string hex)
-        {
-            var lengthPadded = (hex.Length / 2) + (hex.Length % 2);
-            var bytes = new byte[lengthPadded];
-            if (hex == null)
-            {
-                return null;
-            }
-
-            var temp = hex.PadLeft(lengthPadded * 2, '0'); //*2 since a byte is two characters
-
-            var j = 0;
-            for (var i = 0; i < lengthPadded * 2; i += 2)
-            {
-                try
-                {
-                    var chars = temp.Substring(i, 2);
-
-                    bytes[j] = (byte)Convert.ToUInt32(chars, 16);
-                }
-                catch (FormatException e)
-                {
-                    Console.Write(e);
-                    return null;
-                }
-
-                j++;
-            }
-            return bytes;
-        }
-
-        /// <summary>
-        /// Gets you a byte array from the CONTENTS of a string
-        /// </summary>
-        /// <param name="hex"></param>
-        /// <returns></returns>
-        public static byte?[] StringToNullableByteArray(string hex)
-        {
-            var lengthPadded = (hex.Length / 2) + (hex.Length % 2);
-            var bytes = new byte?[lengthPadded];
-            if (hex == null)
-            {
-                return null;
-            }
-
-            var temp = hex.PadLeft(lengthPadded * 2, '0'); //*2 since a byte is two characters
-
-            var j = 0;
-            for (var i = 0; i < lengthPadded * 2; i += 2)
-            {
-                try
-                {
-                    var chars = temp.Substring(i, 2);
-
-                    if (chars == "??")
-                        bytes[j] = null;
-                    else
-                        bytes[j] = (byte)Convert.ToUInt32(chars, 16);
-                }
-                catch (FormatException e)
-                {
-                    Console.Write(e);
-                    return null;
-                }
-
-                j++;
-            }
-            return bytes;
-        }
-
-
-        /// <summary>
-        /// Gets you a byte array from the CONTENTS of a string 0 padded on the left to a specific length
-        /// </summary>
-        /// <param name="hex"></param>
-        /// <returns></returns>
-        public static byte[] StringToByteArrayPadLeft(string hex, int precision)
-        {
-            var bytes = new byte[precision];
-            if (hex == null)
-            {
-                return null;
-            }
-
-            var temp = hex.PadLeft(precision * 2, '0'); //*2 since a byte is two characters
-
-            var j = 0;
-            for (var i = 0; i < precision * 2; i += 2)
-            {
-                try
-                {
-                    if (!byte.TryParse(temp.Substring(i, 2), NumberStyles.HexNumber, CultureInfo.CurrentCulture,
-                        out var b))
-                    {
-                        return null;
-                    }
-
-                    bytes[j] = b;
-                }
-                catch (FormatException e)
-                {
-                    Console.Write(e);
-                    return null;
-                }
-
-                j++;
-            }
-            return bytes;
-        }
-
-        public static string MakeSafeFilename(string filename, char replaceChar)
-        {
-            foreach (var c in Path.GetInvalidFileNameChars())
-            {
-                filename = filename.Replace(c, replaceChar);
-            }
-            return filename;
-        }
-
-        #endregion STRING EXTENSIONS
-
         #region BYTE ARRAY EXTENSIONS
 
         //Thanks to JamieSee https://stackoverflow.com/questions/8440938/c-sharp-left-shift-an-entire-byte-array
